# openapi.yaml
openapi: 3.0.1
info:
  title: Codecademy E-Commerce API
  version: 1.0.0
  termsOfService: https://smartbear.com/terms-of-use/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /products/{id}:
    summary: Gets a specific product
    description: This path gets a specific product (by the id) from the database
    get:
      summary: Gets a specific product
      description: ''
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: 'The id of the product.'
          required: true
          schema:
            type: string
      responses:
        default:
          description: Default error sample response
      tags:
        - Product
  /products/{department}:
    summary: Gets all of the products from a department
    description: >-
      This path gets all of the products in a specific department from the
      database
    get:
      summary: Gets all of the products by department
      description: ''
      operationId: getProductsByDepartment
      parameters:
        - name: department
          in: path
          description: 'The departmant of the product.'
          required: true
          schema:
            type: string
      responses:
        default:
          description: Default error sample response
      tags:
        - Product
...